Add release publishing workflow (#868)
Add a workflow that can be manually triggered to create a new GitHub
release.

The next stage of the automation would be to create a separate workflow
that runs on a schedule, and then triggers this workflow through the
GitHub API if there are relevant changes that need to be released.

Contributes to #846.