Compute next version and add scheduled release (#912)
- Compute the next release version based on changes to components that
are installed in the container. If any installed component has a major
or minor bump, we roll the version to the `x.{y+1}.0` - otherwise we use
`x.y.{z+1}` (i.e. never automatically release a new major version).
- Create a workflow that would automatically create a new release every
Friday at 0900 UTC (or if run manually, principally for test purposes).

The idea is to merge this without the schedule and then me manually run
it see if it correctly detects the changes since the last release in
`docker/` and queues the release workflow to create a draft.

If that works, I'll do a follow-up PR to set `draft:false` and enable
the schedule. Then _in theory_, a release should be automatically
created on the schedule we decide to use.

Contributes to #846.